fix(taro-h5): api以小程序为准scanCode
